Finished up scaling script

+++ data_transformation/data/wiki_information.py
@@ -62,7 +62,7 @@ def parse_scaling(string):
 
 def parse_base_scaling_damage(string):
     print("NOT IMPLEMENTED: " + string)
-    return {"type":"unkown", "value":[]}
+    return {"type":"undefined", "value":[]}
 
 def parse_base_damage(string):
     string = string.replace(u'\u203E',"").replace(u'\xa0', u' ')
@@ -82,9 +82,25 @@ def parse_base_damage(string):
         value = [float(string) for i in range(5)]
     return {"type":type, "value":value}
 
+attr_name_replacements = {
+    "magicdamageperdagger":"magicdamageperinstance",
+    "magicdamagepersphere":"magicdamageperinstance",
+    "magicdamagepermine":"magicdamageperinstance",
+    "physicaldamageperblade": "physicaldamageperinstance",
+    "damageperarrow":"damageperinstance",
+    "magicdamageevery0.25seconds":"magicdamageperquarterseconds",
+    "enhancedmagicdamageevery0.25seconds":"enhancedmagicdamageperquarterseconds"
+}
+def sanitize_attribute_name(attr_name):
+    string = attr_name.replace(" ", "").replace(":","")
+    string = string.replace("min.","minimum").replace("max.","maximum")
+    string.strip()
+    string = attr_name_replacements.get(string, string)
+    return string
+
 def parse_skill(skill):
     if len(skill.get("class")) != 2:
-        return {"type":"unkown"}
+        return {"type":"undefined"}
     info = {"type":skill.get('class',[])[-1], "attributes":{}}
     wrapper = skill.select(".skill_header > div")[0]
     info["name"] = wrapper.get("id")
@@ -92,7 +108,7 @@ def parse_skill(skill):
         # Ignore passives for now
         for level in skill.select(".skill-tabs"):
             for i in range(int(len(level) / 2)):
-                title = html_to_string(level.select("dt:nth-of-type(%d) > span" % (i+1))[0])
+                title = html_to_string(level.select("dt:nth-of-type(%d) > span" % (i+1))[0]).lower()
                 scaling = level.select("dd:nth-of-type(%d)" % (i+1))[0]
                 scales = []
                 for span in scaling.findChildren():
@@ -113,10 +129,10 @@ def parse_skill(skill):
                 if len(base_text) > 0:
                     basedamage = parse_base_damage(base_text)
                     basedamage["scalings"] = scales
-                    info["attributes"][title.replace(" ","")] = basedamage
+                    info["attributes"][sanitize_attribute_name(title)] = basedamage
                 else:
                     # There is no base damage (e.g. Ashe Damage per Arrow on Q)
-                    info["attributes"][title.replace(" ", "")] = {"type":"none", "value":[], "scalings":scales}
+                    info["attributes"][sanitize_attribute_name(title)] = {"type":"none", "value":[], "scalings":scales}
     return info
 
 
@@ -128,7 +144,13 @@ def load_champion_information(champ):
     skill_tags = root.select(".skill")
     if "Lee Sin" in root.title.string:
         # Lee Sin has reactivations as extra skills so he needs to be handled differently
-        print("NOT IMPLEMENTED (Lee Sin can't be parsed yet)")
+        flag = False
+        skills = {"name":"default", "skills":{}}
+        for skill in skill_tags:
+            info = parse_skill(skill)
+            skills["skills"][info["type"] + ("_1" if flag else "_0")] = info
+            flag = not flag
+        information["skills"].append(skills)
         pass
     elif len(skill_tags) == 5 or len(skill_tags) == 6:
         # Normal skill amount. No dual form or smth like that
@@ -157,12 +179,68 @@ def load_champion_information(champ):
 
     return information
 
-champlist = get_champion_list()
-print(champlist)
-result = {}
-for champ in champlist:
-    print("Loading " + champ)
-    result[champ] = load_champion_information(champ)
+attributes = {}
 
-with open("wiki.json", "w+") as file:
-    json.dump(result, file, indent=2)
+# Generate scalings.json
+def generate_scalings(wiki_info):
+    scalings = {}
+    for champ, value in wiki_info.items():
+        champ_scaling = {"min":{},"max":{}}
+        # Iterate over skill sets
+        for skills in value["skills"]:
+            # Iterate over every simple skill
+            for key, skill in skills["skills"].items():
+                # Iterate over attributes
+                if "attributes" in skill:
+                    for attr_name, attr in skill["attributes"].items():
+                        if "scalings" not in attr or len(attr["scalings"]) <= 0:
+                            continue
+                        if "total" in attr_name:
+                            continue
+                        if "minimum" in attr_name and attr_name.replace("minimum", "maximum") in skill["attributes"]:
+                            # ignore because it will be handled once the maximum stat show up
+                            continue
+                        if "maximum" in attr_name and attr_name.replace("maximum","minimum") in skill["attributes"]:
+                            minimum = skill["attributes"][attr_name.replace("maximum","minimum")]
+                            average = {"value":[], "type":attr["type"],"scalings":[]}
+                            average_name = attr_name.replace("maximum","average")
+                            for i in range(len(attr["value"])):
+                                average["value"].append((attr["value"][i] + minimum["value"][i]) / 2)
+                            for j in range(len(attr["scalings"])):
+                                scaling = {"type":attr["scalings"][j]["type"], "value":[]}
+                                min_scale = minimum["scalings"][j]
+                                max_scale = attr["scalings"][j]
+                                for k in range(len(max_scale["value"])):
+                                    scaling["value"].append((min_scale["value"][k] + max_scale["value"][k]) / 2)
+                                average["scalings"].append(scaling)
+                            attr = average
+                        if not attr_name in attributes:
+                            attributes[attr_name] = 0
+                        attributes[attr_name] += 1
+                        for s in attr["scalings"]:
+                            if s["type"] == "undefined":
+                                continue
+                            print(s)
+                            # Add scalings onto each other
+                            champ_scaling["min"][s["type"]] = champ_scaling["min"][s["type"]] + s["value"][0] if s["type"] in champ_scaling["min"] else s["value"][0]
+                            champ_scaling["max"][s["type"]] = champ_scaling["max"][s["type"]] + s["value"][-1] if s["type"] in champ_scaling["max"] else s["value"][-1]
+        scalings[champ] = champ_scaling
+    print(sorted(attributes.items(), key=lambda x:x[1]))
+    with open("wiki.json", "w+") as file:
+        json.dump(scalings, file)
+    #print(attributes)
+
+if os.path.isfile("wiki_info.json"):
+    with open("wiki_info.json") as file:
+        generate_scalings(json.load(file))
+else:
+    champlist = get_champion_list()
+    print(champlist)
+    result = {}
+    for champ in champlist:
+        print("Loading " + champ)
+        result[champ] = load_champion_information(champ)
+
+    with open("wiki_info.json", "w+") as file:
+        json.dump(result, file, indent=2)
+    generate_scalings(result)
